fix(trusty-search): reuse correct rfc3339, stamp SHA on full-reindex, async git, scope indexer lock (#1670)
FIX 1 (HIGH): Delete broken hand-rolled Gregorian now_rfc3339 in
reconcile.rs. Make stages::now_rfc3339 (chrono-backed) pub(crate).
Call chrono::Utc::now().to_rfc3339() directly in stamp_handle. Old
code produced month=13 and day=0/32 for months divisible by 30.
Replace weak format-only test with stamp_handle_produces_valid_rfc3339_date
which parses the output and asserts month ∈ 1..=12, day ∈ 1..=31.

FIX 2 (HIGH): No boot-loop risk. Investigation of finish.rs (lines
321-324) confirms spawn_reindex_with_cleanup → finish_reindex already
stamps indexed_head_sha = git::head_sha(root) and last_indexed_at = now
on successful completion. Added an explicit comment at the
trigger_full_reindex call site documenting this.

FIX 3 (HIGH): Convert changed_files_between from sync
std::process::Command to async tokio::process::Command with .output().await
so the Tokio worker thread is not blocked while git runs. Updated all
call sites to .await. tokio/full already enables the process feature.

FIX 4 (LOW): Restructure apply_delta to acquire and drop the
indexer read lock per-file (not once for the entire batch). Guards are
scoped with explicit blocks so each index_file/remove_file call gets
its own lock acquisition. This mirrors the watch_loop::handle_modified
pattern and allows concurrent HTTP reindex requests (write-lock) to
proceed between files.

Split tests into reconcile_tests.rs (1500-SLOC test cap) so
reconcile.rs stays under the 500-SLOC production cap. Line cap: OK.
